ml/training/requirements.txt
# ── NVIDIA GB10 Grace Blackwell (ARM64 / aarch64) ─────────────────────────────
# PyTorch: install nightly for Blackwell (sm_121) support
#   pip install --pre torch torchvision --index-url https://download.pytorch.org/whl/nightly/cu126
#
# onnxruntime-gpu is not available for ARM64+py3.12; use CPU version:
#   pip install onnxruntime
# ─────────────────────────────────────────────────────────────────────────────

torch>=2.12.0.dev0         # Blackwell (sm_121) support in nightly
torchvision>=0.20.0
onnx>=1.17.0
onnxruntime>=1.19.0        # CPU inference (GPU not yet packaged for ARM64/py3.12)
pandas>=2.0.0
pillow>=10.0.0
tqdm>=4.66.0
tensorboard>=2.15.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
numpy>=1.26.0
